+++ payu/controllers/front/success.php
@@ -3,7 +3,7 @@
  * PayU success
  *
  * @author    PayU
- * @copyright Copyright (c) 2014 PayU
+ * @copyright Copyright (c) PayU
  * @license   http://opensource.org/licenses/LGPL-3.0  Open Software License (LGPL 3.0)
  *
  * http://www.payu.com
@@ -13,38 +13,111 @@
 
 class PayUSuccessModuleFrontController extends ModuleFrontController
 {
+
+    private $order;
+
     public function initContent()
     {
+        parent::initContent();
+
         $payu = new PayU();
 
-        $payu->id_cart = Tools::getValue('id_cart');
+        $order_payment = $payu->getOrderPaymentByExtOrderId(Tools::getValue('id'));
 
-        if (Tools::getValue('error')) {
-            $opc = (bool)Configuration::get('PS_ORDER_PROCESS_TYPE');
-            Tools::redirect('order' . ($opc ? '-opc' : '') . '.php?error=' . Tools::getValue('error'), __PS_BASE_URI__, null, 'HTTP/1.1 301 Moved Permanently');
+        if (!$order_payment) {
+            Tools::redirect('index.php?controller=history', __PS_BASE_URI__, null, 'HTTP/1.1 301 Moved Permanently');
         }
 
-        $payu->payu_order_id = $this->context->cookie->__get('payu_order_id');
+        $payu->id_order = $order_payment['id_order'];
+        $payu->id_cart = $order_payment['id_cart'];
+        $payu->payu_order_id = $order_payment['id_session'];
 
-        $order_payment = $payu->getOrderPaymentBySessionId($payu->payu_order_id);
+        $payu->updateOrderData();
 
-        if ($order_payment) {
-            $payu->id_order = (int)$order_payment['id_order'];
-            $payu->id_cart = (int)$order_payment['id_cart'];
+        $this->order = new Order($payu->id_order);
+        $currentState = $this->order->getCurrentStateFull($this->context->language->id);
 
-            if (Cart::isGuestCartByCartId($payu->id_cart)) {
-                $order = new Order($payu->id_order);
-                $customer = new Customer((int)$order->id_customer);
-                $redirectLink = 'index.php?controller=guest-tracking&id_order=' . $order->reference . '&email=' . urlencode($customer->email);
-            } else {
-                $redirectLink = 'index.php?controller=order-detail&id_order=' . $payu->id_order;
-            }
+        $this->context->smarty->assign(array(
+            'payuLogo' => $payu->getPayuLogo(),
+            'orderPublicId' => $this->order->getUniqReference(),
+            'redirectUrl' => $this->getRedirectLink($payu->id_cart, $payu->id_order),
+            'orderStatus' => $currentState['name'],
+            'HOOK_ORDER_CONFIRMATION' => $this->displayOrderConfirmation(),
+            'HOOK_PAYMENT_RETURN' => $this->displayPaymentReturn()
+        ));
 
-            $payu->updateOrderData();
-            Tools::redirect($redirectLink, __PS_BASE_URI__, null, 'HTTP/1.1 301 Moved Permanently');
-        }
 
-        Tools::redirect('index.php?controller=history', __PS_BASE_URI__, null, 'HTTP/1.1 301 Moved Permanently');
+
+        $this->setTemplate($payu->buildTemplatePath('status', 'front'));
 
     }
+
+    private function getRedirectLink($id_cart, $id_order)
+    {
+        if (Cart::isGuestCartByCartId($id_cart)) {
+
+            $customer = new Customer((int)$this->order->id_customer);
+
+            return $this->context->link->getPageLink(
+                'guest-tracking',
+                null,
+                $this->context->language->id,
+                ['id_order' => $this->order->reference, 'email' => urlencode($customer->email)]
+            );
+        }
+
+        return $this->context->link->getPageLink(
+            'order-detail',
+            null,
+            $this->context->language->id,
+            ['id_order' => $id_order]
+        );
+    }
+
+    /**
+     * Execute the hook displayPaymentReturn
+     */
+    private function displayPaymentReturn()
+    {
+        $params = $this->displayHook();
+
+        if ($params && is_array($params)) {
+            return Hook::exec('displayPaymentReturn', $params, $this->module->id);
+        }
+
+        return false;
+    }
+
+    /**
+     * Execute the hook displayOrderConfirmation
+     */
+    private function displayOrderConfirmation()
+    {
+        $params = $this->displayHook();
+
+        if ($params && is_array($params)) {
+            return Hook::exec('displayOrderConfirmation', $params);
+        }
+
+        return false;
+    }
+
+    private function displayHook()
+    {
+        if (Validate::isLoadedObject($this->order)) {
+            $currency = new Currency((int) $this->order->id_currency);
+
+            $params = array();
+            $params['objOrder'] = $this->order;
+            $params['currencyObj'] = $currency;
+            $params['currency'] = $currency->sign;
+            $params['total_to_pay'] = $this->order->getOrdersTotalPaid();
+
+            return $params;
+        }
+
+        return false;
+    }
+
+
 }